#55: Change Half_up to half_even, added tests

+++ src/main/java/com/github/bgora/rpnlibrary/advanced/functions/CosinusFunctionStrategy.java
@@ -34,8 +34,8 @@ public class CosinusFunctionStrategy extends AbstractFunctionStrategy {
 
     @Override
     public BigDecimal execute(String... params) {
-        BigDecimal param = new BigDecimal(params[0]);
-        BigDecimal result = BigDecimal.valueOf(Math.cos(param.doubleValue()));
+        Double param = new Double(params[0]);
+        BigDecimal result = BigDecimal.valueOf(Math.cos(param));
         return result;
     }
